Chunk Meshes are now showing up. However the coordinates seem to be wrong and 50% of the chunks won't show up

+++ Source/MC_Fake/World/Chunk.h
@@ -5,7 +5,7 @@
 #include "Blocks/Block.h"
 
 struct ChunkGenMaps {
-	TArray<TArray<int32>> HeightMap;
+	TArray<TArray<float>> HeightMap;
 	TArray<TArray<float>> TemperatureMap;
 	TArray<TArray<float>> PrecipitationMap;
 	//River & etc. data
